Utils.ConfigFile: move parsing code into Internal namespace
